requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
beautifulsoup4==4.9.0
cairocffi==1.2.0
CairoSVG==2.3.0
cffi==1.14.4
cssselect2==0.2.1
Cython==0.29.23
defusedxml==0.5.0
dj-database-url==0.5.0
Django==3.2.5
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-cors-headers==3.7.0
django-crispy-forms==1.9.0
django-easy-pdf==0.1.1
django-graphql-jwt==0.3.3
django-heroku==0.3.1
future==0.18.2
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
html5lib==1.0.1
iso8601==1.0.0
Markdown==3.1.1
numpy==1.21.4
pandas==1.3.5
Pillow==8.1.0
promise==2.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.19
PyJWT==2.1.0
PyPDF2==1.26.0
Pyphen==0.9.5
python-bidi==0.4.2
python-dateutil==2.7.5
pytz==2018.7
reportlab==3.6.2
rjsmin==1.1.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
soupsieve==2.0
sqlparse==0.4.1
tinycss2==0.6.1
WeasyPrint==45
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.3